backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Workout Tracker API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlmodel>=0.0.24",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.28",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]
